requirements.txt
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_iam